Move SHM backing from frontend_wayland to core server
This is code that is useful anywhere we need to handle mmap-backed shared memory;
the pressing need is for `wl_shm` buffers, but eg: keymap handling also requires
passing `mmap`able file descriptors over the wire.
